check_everything_connected: assume alternate ref tips are valid
When we receive a remote ref update to sha1 "X", we want to
check that we have all of the objects needed by "X". We can
assume that our repository is not currently corrupted, and
therefore if we have a ref pointing at "Y", we have all of
its objects.  We can therefore stop our traversal from "X"
as soon as we hit "Y".

If we make the same non-corruption assumption about any
repositories we use to store alternates, then we can also
use their ref tips to shorten the traversal.

This is especially useful when cloning with "--reference",
as we otherwise do not have any local refs to check against,
and have to traverse the whole history, even though the
other side may have sent us few or no objects.