t9350-fast-export: Add failing test for symlink-to-directory
git fast-export | git fast-import fails to preserve a commit that replaces
a symlink with a directory.  Add a failing test case demonstrating this
bug.

The fast-export output for the commit in question looks like

  commit refs/heads/master
  mark :4
  author …
  committer …
  data 4
  two
  M 100644 :1 foo/world
  D foo

fast-import deletes the symlink foo and ignores foo/world.  Swapping the M
line with the D line would give the correct result.

Signed-off-by: Anders Kaseorg <andersk@mit.edu>