index-pack: summarily reject non-blobs over warn-object-size
The point of this option being "warn" and not a strict
maximum is that we want to write a nicer message to the
user. We get the complete list of large objects, and then we
walk the rejected history to show their path names.

This is a good idea for blobs, where people accidentally add
large files, and where the cost of walking the history is
low (because we don't really do anything with blobs except
write them to disk, there's little harm in things like
overflows).

But the equation is different for other objects, like trees.
Anybody with a 100MB tree is almost certainly doing
something malicious. And we are at much higher risk, because
we actually parse those trees.

For example, the is_ntfs_dotgit() function (itself a safety
check!) is run even on large trees, but can experience
integer overflow with a 2-gigabyte pathname. This shouldn't
be an exploitable vulnerability (we end up reading NULL, but
the function does not write at all), but it still causes us
to segfault.

Instead, let's summarily die when we see a large object that
isn't a blob.