am: simplify allocations in get_commit_info()
After we call split_ident_line(), we have several begin/end
pairs for bits of the ident. We then copy them into a strbuf
to create a single string, and then duplicate that string.
We can instead skip the strbuf entirely and just duplicate
the strings directly.

This is shorter, and it makes it more obvious that we are
not leaking the strbuf (we were not before, because every
code path either died or hit a strbuf_detach).