status: fix missing newline when comment chars are disabled
When git-status shows tracking data for the current branch
in the long format, we try to end the stanza with a blank
line. When status.displayCommentPrefix is true, we call
color_fprintf_ln() to do so. But when it's false, we call
the enigmatic:

  fputs("", s->fp);

which does nothing at all! This is a bug from 7d7d68022
(silence a bunch of format-zero-length warnings,
2014-05-04). Prior to that, we called fprintf_ln() with an
empty string. Switching to fputs() meant we needed to
include the "newline in the string, but we didn't.

The fix itself is a one-liner. But we never noticed this
because t7508 doesn't exercise the ahead/behind code at all.
So let's configure an upstream during the initial setup,
which means that the code will be exercised as part of all
of the various invocations in that script. This makes the
diff rather noisy, but should give us good coverage.